pyproject.toml
[project]
name = "data-engineering-zoomcamp-26"
version = "0.1.0"
description = "Data Engineering Zoomcamp"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=22.0.0",
    "sqlalchemy>=2.0.45",
    "tqdm>=4.67.1",
    "wget>=3.2",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "pgcli>=4.4.0",
]
